+++ lib/src/audio/audio_responses.dart
@@ -16,11 +16,11 @@ class AudioSuccess extends AudioResponse {
 }
 
 class AudioFailedBadRequest extends AudioResponse {
-  AudioFailedBadRequest()
+  AudioFailedBadRequest({String? reasonPhrase})
       : super(
             code: 400,
             reason:
-                "Bad Request	A required parameter is missing, empty, or null. Or, the value passed to either a required or optional parameter is invalid. A common issue is a header that is too long");
+                "Bad Request	A required parameter is missing, empty, or null. Or, the value passed to either a required or optional parameter is invalid. A common issue is a header that is too long. ${reasonPhrase ?? ''}");
 }
 
 class AudioFailedUnauthorized extends AudioResponse {